notifications/src/requirements.txt
aiohttp==3.8.5
APScheduler==3.10.4
pydantic[email]==1.9.1
uvicorn==0.23.2
python-dotenv==1.0
fastapi==0.100.1
orjson==3.9.5
gunicorn==21.2.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.6.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
asyncpg==0.28.0
alembic==1.11.3
aio-pika==9.2.2
sentry-sdk==1.29.2
python-logstash==0.4.8
sendgrid==6.10.0
Jinja2==3.1.2
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-instrumentation-redis==0.38b0
opentelemetry-instrumentation-sqlalchemy==0.38b0
opentelemetry-exporter-jaeger==1.17.0
asgi-correlation-id==4.2.0
starlette_context==0.3.6